+++ src/index.ts
@@ -1,0 +1,12 @@
+import type { Platform, Arch } from "./platform";
+
+import { osArchPair } from "./platform";
+import BinBuilder from "./builder";
+
+export {
+    type Platform,
+    type Arch,
+
+    osArchPair,
+};
+export default BinBuilder;